+++ src/components/SavePage.jsx
@@ -2,12 +2,12 @@ import React, { useState, useEffect } from 'react';
 import axios from 'axios';
 import PropTypes from 'prop-types';
 
-import Options from './Options';
+import Settings from './Settings';
 
 const SavePage = (props) => {
   const { accessToken } = props;
   const [saveStatusMessage, setSaveStatusMessage] = useState('saving...');
-  const [openOptionsButtonClicked, setOpenOptionsButtonClicked] = useState(false);
+  const [openSettingsButtonClicked, setOpenSettingsButtonClicked] = useState(false);
 
   const api = 'http://localhost:8000/api/v1/pages';
   const headers = { Authorization: accessToken };
@@ -34,14 +34,14 @@ const SavePage = (props) => {
 
   return (
     <>
-      {!openOptionsButtonClicked && (
+      {!openSettingsButtonClicked && (
         <>
           <p>{saveStatusMessage}</p>
-          <button onClick={() => setOpenOptionsButtonClicked(true)}>オプション</button>
+          <button onClick={() => setOpenSettingsButtonClicked(true)}>オプション</button>
         </>
       )}
-      {openOptionsButtonClicked && (
-        <Options />
+      {openSettingsButtonClicked && (
+        <Settings />
       )}
     </>
   );
